Sync changes on designated roles to the internal mongo database
A designated "patron" role can be added by 3rd party
discord bot (donatebot) or by server administrators.

When a Discord members marked as a "patron" on the server,
that means the member gained some perks for sponsoring the
community.

In the first part, we will use it for automatic curation. This changeset
only includes the synchronization part.